+++ boilerplates/aws/src/app/page.tsx
@@ -1,5 +1,4 @@
 import GreetingForm from '@/app/GreetingForm'
-import Image from 'next/image'
 
 export default function Home() {
   return (