+++ boilerplates/aws/src/database/migrations/1729350679372_create-myapp-schema.ts
@@ -1,0 +1,18 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+import { MigrationBuilder, ColumnDefinitions, PgLiteral } from 'node-pg-migrate'
+
+export const shorthands: ColumnDefinitions | undefined = {
+  timestamp: {
+    type: 'timestamptz',
+    notNull: true,
+    default: PgLiteral.create('current_timestamp'),
+  },
+}
+
+export async function up(pgm: MigrationBuilder): Promise<void> {
+  pgm.createSchema('myapp', { ifNotExists: true })
+}
+
+export async function down(pgm: MigrationBuilder): Promise<void> {
+  pgm.dropSchema('myapp', { cascade: true })
+}