+++ packages/db/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nextastic/db
 
+## 1.2.1
+
+### Patch Changes
+
+- add aws migrations and fix migration path
+
 ## 1.2.0
 
 ### Minor Changes